work on multicore scheduling

+++ headers/spinlock.h
@@ -3,5 +3,5 @@
 
 #include "types.h"
 
-u8 acquire(u64 *lock_addr);
-u8 release(u64 *lock_addr);
+u8 acquire(u8 *lock_addr);
+u8 release(u8 *lock_addr);
